move FolderStrategyType to right package
