+++ src/main/java/de/marhali/easyi18n/settings/ProjectSettingsComponentState.java
@@ -4,7 +4,7 @@ import com.intellij.openapi.ui.ComboBox;
 import com.intellij.openapi.ui.TextFieldWithBrowseButton;
 
 import de.marhali.easyi18n.io.parser.ParserStrategyType;
-import de.marhali.easyi18n.model.FolderStrategyType;
+import de.marhali.easyi18n.io.folder.FolderStrategyType;
 import de.marhali.easyi18n.settings.presets.Preset;
 
 import javax.swing.*;
