+++ src/main/java/de/marhali/easyi18n/settings/ProjectSettingsState.java
@@ -1,7 +1,7 @@
 package de.marhali.easyi18n.settings;
 
 import de.marhali.easyi18n.io.parser.ParserStrategyType;
-import de.marhali.easyi18n.model.FolderStrategyType;
+import de.marhali.easyi18n.io.folder.FolderStrategyType;
 import de.marhali.easyi18n.settings.presets.DefaultPreset;
 
 import org.jetbrains.annotations.NotNull;
